Automatically clear the clipboard
Automatically clear the system keyboard after 90 seconds. This value
is configurable. The clipboard is only cleared if it still contains
the content which was originally copied.
To not leak any valueable information about the copied contents,
the clearing background process only receives a truncated hash of
the data which he compares against the (possibly) altered clipboard
contents. Only if they match, the clipboard is wiped.