Allow overwriting the path to a safe file
This change adds a new persistend command line flag, "--safe",
which can be used to temporarily overwrite the path to a pick safe
file as defined in the pick config (or the default path if no such
config file exists).
In addition, load the config from file via "viper" which makes
flag handling and config loading a lot nicer.